# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Job Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface UserProfile {
  id: number;
  currentTitle: string;
  skills: string[];
  yearsExperience: number;
  industry: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  desiredCompMin?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface UserProfileInput {
  /** @minLength 1 */
  currentTitle: string;
  /** @minItems 1 */
  skills: string[];
  /** @minimum 0 */
  yearsExperience: number;
  /** @minLength 1 */
  industry: string;
  location?: string;
  linkedinUrl?: string;
  bio?: string;
  desiredCompMin?: number;
}

export interface TargetRole {
  id: number;
  roleName: string;
  location: string;
  compFloor: number;
  /** @nullable */
  priority?: number | null;
  createdAt: string;
}

export interface TargetRoleInput {
  /** @minLength 1 */
  roleName: string;
  /** @minLength 1 */
  location: string;
  /** @minimum 0 */
  compFloor: number;
  priority?: number;
}

export interface JobPosting {
  id: string;
  title: string;
  company: string;
  location: string;
  /** @nullable */
  salary: number | null;
  /** @nullable */
  salaryMax?: number | null;
  description: string;
  postedAt: string;
  source: string;
  url: string;
}

export interface MatchedJob {
  job: JobPosting;
  /** Percentage 0-100 */
  matchScore: number;
  matchedSkills: string[];
  missingSkills: string[];
}

export interface JobSearchResult {
  jobs: JobPosting[];
  total: number;
  page: number;
}

export interface SkillGapItem {
  skill: string;
  inDemandCount: number;
  hasSkill: boolean;
  importance?: string;
}

export interface SkillAnalysis {
  targetRoleId: number;
  roleName: string;
  matchPercentage: number;
  matchedSkills: string[];
  gapSkills: SkillGapItem[];
  totalJobsAnalyzed: number;
}

export interface CompBand {
  p25: number;
  p50: number;
  p75: number;
  p90: number;
}

export interface CompData {
  role: string;
  location: string;
  band: CompBand;
  /** User's estimated percentile 0-100 */
  userPercentile: number;
  /** up, down, or flat */
  trend: string;
  /** Percentage change vs last quarter */
  trendPercent: number;
  sampleSize: number;
}

export interface CareerStep {
  title: string;
  typicalYearsToReach: number;
  salaryRange: string;
  skillsNeeded: string[];
  /** Percentage of similar profiles who made this transition */
  transitionRate: number;
}

export interface CareerTrajectory {
  currentTitle: string;
  nextRoles: CareerStep[];
  lateralMoves: CareerStep[];
  insights: string[];
}

export interface PeerBenchmark {
  targetRole: string;
  skillPercentile: number;
  experiencePercentile: number;
  overallPercentile: number;
  sampleSize: number;
  insights: string[];
}

export interface Alert {
  id: number;
  /** job_match, comp_change, trajectory, skill_gap */
  type: string;
  message: string;
  isRead: boolean;
  /** @nullable */
  targetRole?: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  newJobMatchesThisWeek: number;
  topMatchScore: number;
  unreadAlerts: number;
  activeTargetRoles: number;
  profileComplete: boolean;
  /** @nullable */
  topTargetRole: string | null;
  /** @nullable */
  avgCompPercentile?: number | null;
  /** @nullable */
  peakMatchRole?: string | null;
}

export type SearchJobsParams = {
role: string;
location: string;
page?: number;
};

export type GetJobMatchesParams = {
targetRoleId?: number;
minMatch?: number;
};

export type GetCompDataParams = {
role?: string;
location?: string;
};