# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Job Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: number;
  /** job_match, comp_change, trajectory, skill_gap */
  type: string;
  message: string;
  isRead: boolean;
  /** @nullable */
  targetRole?: string | null;
  createdAt: string;
}