# lib/api-zod/src/generated/types/careerTrajectory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Job Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CareerStep } from './careerStep';

export interface CareerTrajectory {
  currentTitle: string;
  nextRoles: CareerStep[];
  lateralMoves: CareerStep[];
  insights: string[];
}