# lib/api-zod/src/generated/types/compBand.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Job Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CompBand {
  p25: number;
  p50: number;
  p75: number;
  p90: number;
}