# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Job Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  newJobMatchesThisWeek: number;
  topMatchScore: number;
  unreadAlerts: number;
  activeTargetRoles: number;
  profileComplete: boolean;
  /** @nullable */
  topTargetRole: string | null;
  /** @nullable */
  avgCompPercentile?: number | null;
  /** @nullable */
  peakMatchRole?: string | null;
}