# lib/api-zod/src/generated/types/jobPosting.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Job Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface JobPosting {
  id: string;
  title: string;
  company: string;
  location: string;
  /** @nullable */
  salary: number | null;
  /** @nullable */
  salaryMax?: number | null;
  description: string;
  postedAt: string;
  source: string;
  url: string;
}