# lib/api-zod/src/generated/types/skillGapItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Job Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SkillGapItem {
  skill: string;
  inDemandCount: number;
  hasSkill: boolean;
  importance?: string;
}