# lib/api-zod/src/generated/types/targetRole.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Job Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface TargetRole {
  id: number;
  roleName: string;
  location: string;
  compFloor: number;
  /** @nullable */
  priority?: number | null;
  createdAt: string;
}