# lib/api-zod/src/generated/types/userProfileInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Job Intelligence Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileInput {
  /** @minLength 1 */
  currentTitle: string;
  /** @minItems 1 */
  skills: string[];
  /** @minimum 0 */
  yearsExperience: number;
  /** @minLength 1 */
  industry: string;
  location?: string;
  linkedinUrl?: string;
  bio?: string;
  desiredCompMin?: number;
}